add StaticResponseFormatter for serving static files
