+++ CityWebServer.Extensibility/CityWebServer.Extensibility.csproj
@@ -30,6 +30,9 @@
     <WarningLevel>4</WarningLevel>
   </PropertyGroup>
   <ItemGroup>
+    <Reference Include="ColossalManaged">
+      <HintPath>..\ColossalManaged.dll</HintPath>
+    </Reference>
     <Reference Include="JsonFx">
       <HintPath>..\packages\JsonFx.2.0.1209.2802\lib\net35\JsonFx.dll</HintPath>
     </Reference>
@@ -39,8 +42,12 @@
     <Reference Include="System.Data.DataSetExtensions" />
     <Reference Include="System.Data" />
     <Reference Include="System.Xml" />
+    <Reference Include="UnityEngine">
+      <HintPath>..\UnityEngine.dll</HintPath>
+    </Reference>
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="ApacheMimeTypes.cs" />
     <Compile Include="IWebServer.cs" />
     <Compile Include="RequestHandlerBase.cs" />
     <Compile Include="ResponseFormatters\HtmlResponseFormatter.cs" />
@@ -51,6 +58,7 @@
     <Compile Include="LogAppenderEventArgs.cs" />
     <Compile Include="ResponseFormatters\PlainTextResponseFormatter.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="ResponseFormatters\StaticResponseFormatter.cs" />
     <Compile Include="RestfulRequestHandlerBase.cs" />
   </ItemGroup>
   <ItemGroup>
