+++ CityWebServer/CityWebServer.csproj
@@ -61,7 +61,6 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Helpers\ApacheMimeTypes.cs" />
     <Compile Include="Helpers\CitizenExtensions.cs" />
     <Compile Include="Helpers\ConfigurationHelper.cs" />
     <Compile Include="Helpers\DistrictExtensions.cs" />
